% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tab_exp.R
\name{base_tab_exp}
\alias{base_tab_exp}
\title{Slight formatting for exporting a base tab object}
\usage{
base_tab_exp(object, replacer = " -:- ")
}
\arguments{
\item{object}{A ucr.base.tab object or the 'tab' component of that list.}

\item{replacer}{Replace the LaTeX in first column with this}
}
\description{
Replace LaTeX code with something more readable. This
    is a simple convenience function that gsub:s the latex code
    from the first row of the ucr base tab 'tab' component.
}
\examples{
## Apply this to the base tab object to make it more readable
## bt <- ucr.base.tab(my_data)
## write.csv(ucr_tab_exp(bt), file = "path/to/file.csv")
}
\author{
Henrik Renlund
}

