% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tab_prep.r
\name{base_tab_prep}
\alias{base_tab_prep}
\title{Prepare a data frame for base tabulating}
\usage{
base_tab_prep(X, max.unique = 10, elim.names = NULL, force.factor = TRUE,
  binary.code = TRUE, elim.class = NULL)
}
\arguments{
\item{X}{a data frame}

\item{max.unique}{maximum number of unique values to tolerate (for
non-numeric variables), default 10}

\item{elim.names}{set of names to eliminate}

\item{force.factor}{\code{TRUE} (default) if caracter vectors should be
factors}

\item{binary.code}{convert a 0/1 variable into a factor with these labels (if
\code{TRUE} then 0 = "No" and 1 = "Yes").}

\item{elim.class}{a set of classes to eliminate in addition to \code{Date},
\code{POSIXct} and \code{POSIXt}}
}
\description{
Remove difficult variables: dates, factors with too many unique
  values, etc.
}
\author{
Henrik Renlund
}
