% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ucr_model_tab_methods.r
\name{cbind.ucr.model.tab}
\alias{cbind.ucr.model.tab}
\title{Combine two or more model tables.}
\usage{
cbind.ucr.model.tab(...)
}
\arguments{
\item{...}{a number of \code{ucr.model.tab} objects.}
}
\value{
Return a new \code{ucr.model.tab} object with more columns.
}
\description{
Combine two or more model tables. The tables are displayes
             in separate columns. The tables are supposed to be identical
             except for the regression results.
}
\author{
Lars Lindhagen
}
\seealso{
\code{\link{ucr.model.tab}}
}

