% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confint_coxme.r
\name{confint.coxme}
\alias{confint.coxme}
\title{\code{confint} method for class \code{coxme}}
\usage{
\method{confint}{coxme}(object)
}
\arguments{
\item{object}{an onbject of class \code{coxme}}
}
\description{
This is a crappy solution to getting the confidence intervals
for the coefficients from a \code{coxme} object. It captures the output from
\code{print(object)}... Oh, well.
}
\author{
Henrik Renlund
}

