% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cr_group.r
\name{cr_group}
\alias{cr_group}
\title{Create \code{cr_group} object}
\usage{
cr_group(x, rgroup, cgroup, colnames)
}
\arguments{
\item{x}{the object (typically a matrix och data frame)}

\item{rgroup}{the 'rgroup' attribute with length equal to \code{dim(object)[1]}}

\item{cgroup}{the 'cgroup' attribute with length equal to \code{dim(object)[2]}}

\item{colnames}{use this to set colnames (class \code{cr_group} can keep track of these under permutations (if object is 'data.frame' colnames are dropped under permuation)}
}
\description{
Sets the 'rgroup' and 'cgroup' attributes and returns a
'cr_group' object. These can also be set with \code{attr}, but this
functions checks that the dimensionality is OK.
}
\note{
This function will be placed in the 'dataman' package.
(It will be eventually be removed so as to not clutter the 'ucR' package too
much, but I don't want to break to much code in current use!)
}
\author{
Henrik Renlund
}

