% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_man_functions.r
\name{data_man}
\alias{data_man}
\title{Data Management Tool (interactive-ish)}
\usage{
data_man(name, var = NA, where = NA, recode = NULL, date = NULL,
  rg = NULL, comment = NULL, container = "data_man_container",
  label = FALSE, check = TRUE)
}
\arguments{
\item{name}{(new) name of variable}

\item{var}{database entry}

\item{where}{database (or data frame)}

\item{recode}{ucR::reFactor L argument for recoding}

\item{date}{function for transformation into date format}

\item{rg}{rgroup label}

\item{comment}{variable comment}

\item{container}{name of container of information in \code{.GlobalEnv}.
Defaults to "data_man_container"}

\item{label}{if TRUE then the data base entry will be stored as a label for
the new variable. Another label can be specified with a character string.}

\item{check}{if TRUE this checks the variable}
}
\description{
...
}
\author{
Henrik Renlund
}

