% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_man_functions.r
\name{data_man_create}
\alias{data_man_create}
\title{Get data from \code{data_man_container} info}
\usage{
data_man_create(id, container, id.name, na.where = NULL)
}
\arguments{
\item{id}{vector of id's}

\item{container}{name of \code{data_man_container} object}

\item{id.name}{identifier across data bases or a list of such}

\item{na.where}{if entry \code{where} is \code{NA} in
\code{data_man_container}, this object is used in its place.}
}
\description{
...
}

