% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_base_tab_methods.r
\name{latex.ucr.base.tab}
\alias{latex.ucr.base.tab}
\title{Produce latex code from a \code{ucr.base.tab} object}
\usage{
\method{latex}{ucr.base.tab}(object, alt.gray = F, col.just = NULL, collabel.just = NULL, ...)
}
\arguments{
\item{object}{a \code{ucr.base.tab} object}

\item{alt.gray}{a logical flag telling whether every other row should be gray}

\item{col.just}{if \code{NULL}, left-justification is applied}

\item{collabel.just}{if \code{NULL}, left-justification is applied}

\item{...}{arguments to be passed to \code{latex}}
}
\description{
A simple wrapper for the Hmisc latex function,
that adds an explanatory bottom text and extra
column headings.
}
\seealso{
\code{\link{ucr.base.tab}}
}
\author{
Lars Lindhagen
}
