% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ucr_model_tab_methods.r
\name{latex.ucr.model.tab}
\alias{latex.ucr.model.tab}
\title{Produce latex code from a \code{ucr.model.tab} object}
\usage{
latex.ucr.model.tab(object, ...)
}
\arguments{
\item{object}{a \code{ucr.model.tab} object}

\item{...}{arguments to be passed to \code{latex}}
}
\description{
A simple wrapper for the Hmisc latex function,
that adds extra column headings.
}
\author{
Lars Lindhagen
}
\seealso{
\code{\link{ucr.model.tab}}
}

