% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc-digits.R
\name{magnitude}
\alias{magnitude}
\title{magnitude of number}
\usage{
magnitude(x)
}
\arguments{
\item{x}{a numeric}
}
\description{
returns the magnitude of the input. Why? No reason. Oh, ok, it is actually a
    helper function for \code{trunc_digits}.
}
\author{
Henrik Renlund
}
