% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDF.character}
\alias{makeDF.character}
\title{Turn variables into a data frame}
\usage{
\method{makeDF}{character}(object, env = .GlobalEnv, ...)
}
\arguments{
\item{object}{a character vector of variables contained in \code{env}}

\item{env}{the environment, data frame or list that contains the variables}

\item{...}{for future needs?}
}
\description{
Turn variables into a data frame
}
\author{
Henrik Renlund
}

