% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDF.list}
\alias{makeDF.list}
\title{Turn elements of a list into a data frame}
\usage{
\method{makeDF}{list}(object, env = .GlobalEnv, ...)
}
\arguments{
\item{object}{a list of objects for which a makeDF method exist}

\item{env}{environment must currently be set to \code{.GlobalEnv}}

\item{...}{for future needs?}
}
\description{
Turn elements of a list into a data frame
}
\author{
Henrik Renlund
}

