% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDF}
\alias{makeDF}
\title{Turn objects into data frames}
\usage{
makeDF(object, env, ...)
}
\arguments{
\item{object}{an object}

\item{env}{an environment}

\item{...}{passed arguments}
}
\description{
Turn objects into data frames
}
\author{
Henrik Renlund
}

