% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDF.survfit}
\alias{makeDF.survfit}
\title{Turn survfit object into a data frame}
\usage{
\method{makeDF}{survfit}(object, env = .GlobalEnv, ...)
}
\arguments{
\item{object}{a survfit object}

\item{env}{environment must currently be set to \code{.GlobalEnv}}

\item{...}{for future needs?}
}
\description{
Turn survfit object into a data frame
}
\author{
Henrik Renlund
}

