% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_describe}
\alias{mids_describe}
\title{Describe an imputed data set}
\usage{
mids_describe(object, file = "", ...)
}
\arguments{
\item{object}{a mids object}

\item{file}{file to write latex output to (empty string by default)}

\item{...}{arguments passed to \code{Hmisc::latex}}
}
\description{
Describe complete case, missing and imputed via ucr.base.tab
}
\author{
Henrik Renlund
}

