% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_get}
\alias{mids_get}
\title{Extract imputed data sets}
\usage{
mids_get(object, m)
}
\arguments{
\item{object}{a mids object}

\item{m}{imputation index}
}
\description{
Extract imputed data sets from a \code{mids} object
}
\author{
Henrik Renlund
}

