% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_predict_logreg_2}
\alias{mids_predict_logreg_2}
\title{Predict from imputed data on imputed data}
\usage{
mids_predict_logreg_2(object, formula, newmids)
}
\arguments{
\item{object}{a mids object}

\item{formula}{formula for the logistic regression}

\item{newmids}{data for which predictions are wanted}
}
\description{
Predict from imputed data sets from a \code{mids} object and a
logistic regerssion model
}
\note{
It would be very nice to be able to include \code{se.fit}...
}
\author{
Henrik Renlund
}

