% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_predict_on}
\alias{mids_predict_on}
\title{Predict on imputed data}
\usage{
mids_predict_on(object, mids, ...)
}
\arguments{
\item{object}{prediction object (=model, typically)}

\item{mids}{mids object to be predicted on.}

\item{...}{arguments passed to predict}
}
\description{
Predict from an object (model) on a \code{mids} object.
}
\note{
It would be nice to include the equivalent of parameter \code{se.fit}
from other prediction methods.
}
\author{
Henrik Renlund
}

