% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_subset}
\alias{mids_subset}
\title{Subset imputed data sets}
\usage{
mids_subset(object, index)
}
\arguments{
\item{object}{a mids object}

\item{index}{the rows you want (can be a logical expression)}
}
\description{
Subset imputed data sets from a \code{mids} object
}
\author{
Henrik Renlund
}

