% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_subset_2}
\alias{mids_subset_2}
\title{Subset imputed data sets}
\usage{
mids_subset_2(object, index)
}
\arguments{
\item{object}{a mids object}

\item{index}{the rows you want (must be integer for improper subsets)}
}
\description{
Subset imputed data sets from a \code{mids} object.
If the subsetting index is proper, this function will call \code{mids_subset} and thus \code{index} can be integer, logical or character. For improper subsets the index must be integer, and the \code{object$data} will get its rownames set to \code{1:nrow(object$data)} (for ease of handling).
}
\author{
Henrik Renlund
}

