% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nejm_pallette.R
\name{nejm_palette}
\alias{nejm_palette}
\title{The NEJM palette}
\usage{
nejm_palette(show = FALSE)
}
\arguments{
\item{show}{if true the palette is shown in a new graphical window}
}
\value{
The codes for the eight colors in the NEJM graphical profile
}
\description{
Returns the colors of the NEJM palette
}
\author{
Henrik Renlund
}
