% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ucr_base_tab.R
\name{print.ucr.base.tab}
\alias{print.ucr.base.tab}
\title{Print a ucr.base.tab object}
\usage{
\method{print}{ucr.base.tab}(x, replacer = " -:- ", ...)
}
\arguments{
\item{replacer}{Replace the LaTeX in first column with this}

\item{object}{A ucr.base.tab object or the 'tab' component of that list.}
}
\description{
Format the object with base_tab_exp before printing
}
\author{
Henrik Renlund
}

