% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reFactor}
\alias{reFactor}
\title{This function manipulates factor levels}
\usage{
reFactor(x, L, na.level = NULL, exclude = NULL, new.last = FALSE,
  warn = TRUE)
}
\arguments{
\item{x}{factor or character vector to be relabelled or reordered}

\item{L}{a list specifying the relabeling/reordering}

\item{na.level}{if not \code{NULL} all \code{NA}'s will be a level with this label}

\item{exclude}{a value to make into NA's}

\item{new.last}{if \code{TRUE} then added levels will be placed last in the list of levels (default \code{FALSE})}

\item{warn}{do you want warnings when doing stupid things?}
}
\description{
This functions enables relabeling, reordering and merging of factor levels.
It is also possible to add and exclude levels, and turn missing values into
a separate level.
}
\examples{
x <- LETTERS[1:5] # x <- factor(x)
L <- list(B=c("nyB", "C"),E=NULL, D="nyD")
reFactor(x,L)
}
\author{
Henrik Renlund
}
\seealso{
\code{\link{factor}}, \code{\link{relevel}}
}

