% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spec.R
\name{sens_spec}
\alias{sens_spec}
\title{sensitivity and specificity}
\usage{
sens_spec(status, measure, data = NULL, disease.value = 1, direction = 1)
}
\arguments{
\item{status}{the grouping of diagnosis}

\item{measure}{a predictive value}

\item{data}{a data frame or similar}

\item{disease.value}{if status is not 0/1 specify the value indicating a
positive diagnosis}

\item{direction}{if 1 then a higher value is more indicative of a positive
diagnosis, anything the other way around}
}
\description{
calculate the sensitivity and specificity from a grouping and a measurement
}
