% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match.r
\name{time_match}
\alias{time_match}
\title{find matches on variables in a given time period}
\usage{
time_match(data, s, search, id = "lopnr", date = "INDATUMA", begin = NULL,
  end = NULL, matches = "first.id", non.matches = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{s}{a regular expression}

\item{search}{names of variables to search in (given in order of importance)}

\item{id}{name of id variable}

\item{date}{name of associated date variable}

\item{begin}{date (or vector of dates) or the name of a date variable}

\item{end}{date (or vector of dates) or the name of a date variable}

\item{matches}{specification of what to return; 'all' for all matches,
'first.date' to filter on the first occurence for any individual and
date, and 'first.id' to filter on the first occurence of any individual}

\item{non.matches}{if \code{TRUE} also return data on the individuals with
no matches}
}
\value{
a data frame with \itemize{
 \item id id variable
 \item event 1 of match found (else 0 if any and \code{non.matches = TRUE})
 \item t days from 'begin' to 'date'
 \item match the match found
 \item variable match found in this variable
 \item t2 days from 'date' to end
 \item first.id indicator for first occurence of associted id
 \item first.date indicator for first occurence of associated id and date
 \item begin the begin variable
 \item date the date variable
 \item end the end variable
}
}
\description{
in a dataset with one or more variables (typically containing text or codes)
    associated with a date, find matches on those variables within specifed
    time frames
}
\author{
Henrik Renlund
}
