% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modComp.r
\name{transFormat}
\alias{transFormat}
\title{Transform and format}
\usage{
transFormat(x, fun = NULL, signif = NULL, round = NULL, skip = FALSE,
  ...)
}
\arguments{
\item{x}{the object}

\item{fun}{the function}

\item{signif}{if non-NULL this is the 'digits' argument for \code{signif}}

\item{round}{if non-NULL this is the 'digits' argument for \code{round}}

\item{skip}{if TRUE \code{format} is skipped}

\item{...}{arguments passed to \code{base::format}}
}
\description{
Apply a function to x, then \code{round} or \code{signif} before formatting
}

