% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc-digits.R
\name{trunc_digits}
\alias{trunc_digits}
\title{truncation with digits}
\usage{
trunc_digits(x, digits = 1)
}
\arguments{
\item{x}{a numeric}

\item{digits}{an integer}
}
\description{
like \code{trunc}, but with digits specification
}
\author{
Henrik Renlund
}
