% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ucr_base_tab.r
\name{ucr.base.tab}
\alias{ucr.base.tab}
\title{Create a baseline table}
\usage{
ucr.base.tab(data, group.name = NULL, combined.name = "Combined",
  x.names = setdiff(names(data), group.name), num.format = "median",
  median.format = "iqr", mean.format = "par",
  factor.format = "count.perc", perc.method = "group", print.perc = T,
  print.perc.space = F, omit.ref.level = F, show.missing = "none",
  digits = 1, spec.digits = NULL, include.combined = T, include.n = T,
  include.p = T, num.test = "nonparam", factor.test = "fisher",
  min.p = "0.001")
}
\arguments{
\item{data}{The data frame.}

\item{group.name}{Name of the variable in 'data' that defines the groups, or NULL
(default) if there are no groups. Note that there must be no
NA's in the group variable.}

\item{combined.name}{Column heading for the combined group. Default  "Combined".}

\item{x.names}{Names and order of the variables in 'data' that shall be included in
the table. By default, all variables are included.}

\item{num.format}{How to present numerical variables. Allowed values:
\itemize{
  \item "median": Median with IQR or range.
\item "mean": Mean with standard deviation.
\item "both": Both. This gives "median stuff (mean stuff)".
}}

\item{median.format}{What measure of dispersion to give together with medians.
 Allowed values:
 \itemize{
\item "iqr": Interquartile range (Q1 and Q3).
 \item "range": Range (min and max).
 \item A numerical value strictly between 0 and 50.
   Example: If median.format = 10, then the 10th and 90th
   percentiles will be reported.
   }}

\item{mean.format}{How to present standard deviations. Allowed values:
\itemize{
 \item "par": Mean (SD).
\item "pm": Mean plus/minus SD.
}}

\item{factor.format}{How to present factor variables. Allowed values:
 \itemize{
\item count.perc: Count (percentage).
 \item perc.count: Percentage (count).
 }}

\item{perc.method}{How to compute percentages for factor variables. Allowed values:
\itemize{
 \item "group": Computes n(group, level) / n(group).
\item "level": Computes n(group, level) / n(level).
  For the combined group, this would always give 100%.
  Hence, percentages are omitted in this case.
\item "total": Percentage  n(group, level) / n.
}}

\item{print.perc}{TRUE if a percent sign ('\%') is to be printed for percentages.}

\item{print.perc.space}{TRUE if a space is to be printed between the number and the
percent sign, i.e. "12.3 \%" rather than "12.3\%".
N/A if percent sign is disabled.}

\item{omit.ref.level}{TRUE if the reference (first) level of dichotomous factor
variables should be omitted from the table.}

\item{show.missing}{How should missing values, if any, per group be shown?
\itemize{
\item "none": Don't show missing values.
\item "in.row": Show the number of missing values within square
  brackets after the usual data.
\item "sep.row": Show the number of missing values is a separate
  row.
  }}

\item{digits}{An integer, giving the number of decimals for numeric
variables, unless otherwise specified through the parameter
'spec.digits'.}

\item{spec.digits}{A list, whose names form a subset of 'x.names', and whose values
are the number of decimals for those (numeric) variables. The
value NULL causes 'digits' to be used throughout.}

\item{include.combined}{TRUE if the combined group should be included in a
separate column.}

\item{include.n}{TRUE if number of valid (non-NA) observations per variable should
be included in a separate column.}

\item{include.p}{TRUE if P-values are to be included in the table. They refer to
univariate tests for identical distribution in the different
groups. P-values can only be included if there are groups.}

\item{num.test}{Which kind of test to use for numerical variables. The selected
test depends on the number of groups. Allowed values:
\itemize{
\item "nonparam": A non-parametrical test:
\itemize{
    \item Wilcoxon (Mann-Whitney) test for two groups.
    \item Kruskal-Wallis test for three or more groups.
    }
\item "param": A parametrical test:
    \itemize{
    \item t-test for two groups.
    \item ANOVA for three or more groups.
    }
    }}

\item{factor.test}{Which test to use for factor variables. Allowed values:
\itemize{
 \item "fisher": Fisher's exact test.
\item "pearson": Pearson's chi2 test.
}}

\item{min.p}{Smallest P-value to be displayed. If smaller, then a text like
"< 0.001" is given.
NOTE: This value must be a string, e.g. "0.001" or "1e-3".}
}
\value{
The return value is an S3 object of class ucr.base.tab.
}
\description{
Creates a baseline table from a data frame. The data may
be partitioned in two or more groups. A column for the
"combined" group (all data) will also be included.
}
\author{
Lars Lindhagen
}

