% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_palette.r
\name{ucr_palette}
\alias{ucr_palette}
\title{The UCR palette}
\usage{
ucr_palette(show = FALSE)
}
\arguments{
\item{show}{if true the palette is shown in a new graphical window}
}
\value{
The codes for the five colors in the UCR graphical profile
}
\description{
Returns the colors of the UCR palette
}
\author{
Henrik Renlund
}
