% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confint_glmerMod.r
\name{confint.glmerMod}
\alias{confint.glmerMod}
\title{\code{confint} method for class \code{glmerMod}}
\usage{
\method{confint}{glmerMod}(object, parm = NULL, level = 0.95, ...,
  more = FALSE)
}
\arguments{
\item{object}{an object of class \code{coxme}}

\item{parm}{(only for method compatability)}

\item{level}{(only for method compatability)}

\item{...}{(only for method compatability)}

\item{more}{set to TRUE if you also want the coefficients and p-values.}
}
\description{
This is a highly crappy solution to getting the confidence
  intervals for the coefficients from a \code{coxme} object. It captures the
  output from \code{print(object)}... Oh, well.
}
\author{
Henrik Renlund  - but don't hold it against me
}

