\name{latex.ucr.base.tab}
\alias{latex.ucr.base.tab}
\title{Produce latex code from \code{ucr.base.tab} object}
\usage{
latex.ucr.base.tab(object, ...)
}
\arguments{
  \item{object}{a \code{ucr.base.tab} object}

  \item{...}{arguments to be passed}
}
\description{
A simple wrapper for the Hmisc latex function, that adds an
explanatory bottom text and extra column headings.
}
\author{
Lars Lindhagen
}
\seealso{
\code{\link{ucr.base.tab}}
}

