% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeDF.r
\name{makeDF}
\alias{makeDF}
\title{Turn objects into data frames}
\usage{
makeDF(object, env, ...)
}
\arguments{
\item{object}{an object}

\item{env}{an environment}

\item{...}{passed arguments}
}
\description{
Some R functionality (I'm thinking mainly of ggplot2) requires
the data you are interested in to be packaged in data frames, hence this
function.
}
\note{
This function will be \code{make_df} in the 'dataman' package.
(It will be eventually be removed so as to not clutter the 'ucR' package too
much, but I don't want to break to much code in current use!)
}
\examples{
# combine variables into a data frame
x <- 1:2
y <- letters[1:2]
makeDF(c("x","y"))
# 'env' is R_GlobalEnv by default but can be any environment, data frame or list
makeDF(c("mpg", "disp"), env=datasets::mtcars)
# methods include 'list' ...
makeDF(list(x=1:2, y=letters[1:2]))
# and 'survfit'
library(survival)
n <- 50
s <- Surv(rexp(n), rbinom(n,1,0.2))
f <- factor(rep(letters[1:2], length.out=n))
sf <- survfit(s~f)
makeDF(sf)
}
\author{
Henrik Renlund
}

