% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mids_functions.r
\name{mids_describe}
\alias{mids_describe}
\title{Describe an imputed data set}
\usage{
mids_describe(object, x.names, file = "", ..., force.factor = NULL,
  factorize = TRUE, digits = 1, silent = TRUE)
}
\arguments{
\item{object}{a mids object}

\item{x.names}{character vector of variables of interest}

\item{file}{file to write latex output to (empty string by default)}

\item{...}{arguments passed to \code{Hmisc::latex}}

\item{force.factor}{a character vector of variables to be forced into factors
(e.g. event variables that are coded as 0/1)}

\item{factorize}{force character vectors to be factors? (if FALSE these
variables are removed)}

\item{digits}{the number of digits (median and IQR for imputed
variable description)}

\item{silent}{hide message on removed variables?}
}
\value{
LaTeX code for descriptive table
}
\description{
Describe complete case, missing and imputed via ucr.base.tab
}
\note{
Currently this function utilizes the \code{ucr.base.tab} function with
a certain parameters. It is unclear how easy it would be to allow the
passing of arguments from \code{mids_describe} to this function...
}
\author{
Henrik Renlund
}

