% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match.R
\name{time_match}
\alias{time_match}
\title{find matches for specific units on variables in a given time period}
\usage{
time_match(pattern, x = NULL, data, id = "id", date = "date",
  units = NULL, units.id = id, begin = "begin", end = "end", ...,
  long = TRUE, stack = TRUE, verbose = TRUE)
}
\arguments{
\item{pattern}{a vector of search strings (regular expressions) (the names
attribute will be used if it exists)}

\item{x}{names of variables to search in (given in order of importance), if
missing all variables except id and date are chosen}

\item{data}{a data frame}

\item{id}{name of id variable (in 'data')}

\item{date}{name of associated date variable (in 'data')}

\item{units}{a vector of id's, or a data frame containing id's as well as
(but optionally) 'begin' and 'end' variables}

\item{units.id}{variable name in 'set' to use as id (by default the same is
'id')}

\item{begin}{variable name in 'set' to use as begin, if missing will be set
to earliest date in data}

\item{end}{variable name in 'set' to use as end, if missing will be set to
latest date in data}

\item{...}{arguments passed to \code{grepl}}

\item{long}{if \code{TRUE} all matches will get a row, else first match gets
details and information on all other matches is condensed. N.B
\code{long = FALSE} will be slow for large datasets!}

\item{stack}{if \code{TRUE} results are stacked. Not stacking is only
possible when \code{long = FALSE}.}

\item{verbose}{if \code{TRUE} the function will give helpful and/or annoying
messages}
}
\value{
The basic 'long' output is a data frame with
 \itemize{\item id the id variable
 \item begin the begin date (could be individual)
 \item end the end date (could be individual)
 \item date the date of assicated match
 \item event indicator for a match
 \item time days from 'begin' to 'date'
 \item match the match found
 \item match.in the variable the match was found in
 \item pattern the pattern searched for
 \item alias the name of pattern searched for (else p1, p2, etc)
 \item first.id indicator for first occurence of associated id
 \item first.id_date indicator for first occurence of associated id and date
 \item ... all variables in data that are not id, date or search
    variables. These will be renamed if they are in conflict with output
    names. These will only be included in output when \code{long = TRUE}.}

Note that any individual can have more than one match.

The basic output when \code{long = FALSE} is one line per individual where
    the first match (by date and order of search variables, i.e. filtered on
    \code{first.id == 1}) is specified with some detail, and information on
    all subsequent matches is condensed.

 \itemize{ \item id, begin, end, date, event, time, match, match.in, pattern, alias
    as before, but only relevant for the first match.
 \item events the total number of matches found
 \item matches all (unique) matches found, separated by a space
 \item matches.info all (unique) match/mathing-variable/date-combinations
    separated by a space }

If pattern is a vector, the results can be stacked or not. If they are not
    stacked, the format must necessarily be one line per individual
    (i.e. \code{long = FALSE}).  If the output is unstacked, all pattern
    specific output (i.e. all but id, begin and end) variables will get a
    suffix, either the names of the pattern vector, or a created one.
}
\description{
in a dataset with one or more variables (typically containing text)
    associated with a date, find matches on those variables for specific
    individuals within specifed time frames
}
\author{
Henrik Renlund
}
