% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match.r
\name{time_match}
\alias{time_match}
\alias{time_match_set}
\title{find matches}
\usage{
time_match(data, s, search, id = "id", date = "date", begin = NULL,
  end = NULL, matches = "all", non.matches = TRUE)

time_match_set(data, s, search, id = "id", date = "date", set,
  set.id = "id", set.begin = "begin", set.end = "end", matches = "all",
  non.matches = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{s}{a search string (regular expression)}

\item{search}{names of variables to search in (given in order of importance)}

\item{id}{name of id variable}

\item{date}{name of associated date variable}

\item{begin}{date (or vector of dates) or the name of a date variable}

\item{end}{date (or vector of dates) or the name of a date variable}

\item{matches}{specification of what to return; 'all' for all matches,
'first.date' to filter on the first occurence for any individual and
date, and 'first.id' to filter on the first occurence of any individual}

\item{non.matches}{if \code{TRUE} also return data on the individuals with
no matches}

\item{set}{a vector of id's, or a data frame containing id's as well as (but
optionally) 'begin' and 'end' variables}

\item{set.id}{variable name in 'set' to use as id}

\item{set.begin}{variable name in 'set' to use as begin}

\item{set.end}{variable name in 'set' to use as end}
}
\value{
a data frame with \itemize{
 \item id id variable
 \item event 1 of match found (else 0 if any and \code{non.matches = TRUE})
 \item t days from 'begin' to 'date'
 \item match the match found
 \item variable match found in this variable
 \item t2 days from 'date' to end
 \item first.id indicator for first occurence of associted id
 \item first.date indicator for first occurence of associated id and date
 \item begin the begin variable
 \item date the date variable
 \item end the end variable
}
}
\description{
find matches on variables in a given time period
}
\details{
in a dataset with one or more variables (typically containing text or codes)
    associated with a date, find matches on those variables within specifed
    time frames
}
\section{Functions}{
\itemize{
\item \code{time_match_set}: wrapper for find_match to look at only those units within a given set
}}

\author{
Henrik Renlund
}
