% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucR.r
\docType{package}
\name{ucR}
\alias{ucR}
\title{ucR: miscellaneous functions}
\description{
'ucR' is a package of miscellaneous functions used by the R coders of ucr stat.
It is under git version control at \url{https://github.com/renlund/ucR}.
}
