% Generated by roxygen2 (4.0.2): do not edit by hand
\name{varFind}
\alias{varFind}
\title{Look for a pattern in the \code{names} of R objects (typically data frames) and
return the hits}
\usage{
varFind(pattern, dfs = TRUE, ignore.case = TRUE, ...)
}
\arguments{
\item{pattern}{pattern to look for}

\item{dfs}{logical or character string. If a character vector of the names of
R objects is given, the \code{names} of these will be exmined with \code{grepRet}
; if \code{TRUE} all data frames in the global environment will be examined and if
\code{FALSE} all objects with a 'names' attribute will be examined.}

\item{ignore.case}{logical; should upper/lower case distinction be ignored? (default \code{TRUE})}

\item{...}{arguments to be passed to \code{\link{grep}}}
}
\description{
This function applies \code{\link{grepRet}} to \code{\link{names}} of a set
of R objects and returns the results as a list
}
\examples{
require(datasets)
varFind(pattern="a", dfs=c("mtcars", "esoph"), index=TRUE)
}
\author{
Henrik Renlund
}
\seealso{
\code{\link{grepRet}}, \code{\link{grep}}
}

